fixes stack deletion failure, when vol in deleting
when an delete operation is requested on a stack, and the volume in stack
is already in 'deleting' state, heat fails to delete the stack and goes to
DELETE_FAILED state instantly.

Instead of going to DELETE_FAILED state instantly, heat needs to safely
wait for the volume to complete its deletion process, so that heat can
then go ahead and complete stack deletion successfully.

Closes-Bug: #1362708

+++ heat/engine/resources/volume.py
@@ -160,8 +160,10 @@ class Volume(resource.Resource):
 
                 if vol.status == 'in-use':
                     raise exception.Error(_('Volume in use'))
-
-                vol.delete()
+                # if the volume is already in deleting status,
+                # just wait for the deletion to complete
+                if vol.status != 'deleting':
+                    vol.delete()
                 while True:
                     yield
                     vol.get()
